float
# Conflicts:
#	source/fitz/pixmap.c

+++ source/fitz/pixmap.c
@@ -1132,7 +1132,7 @@ fz_gamma_pixmap(fz_context *ctx, fz_pixmap *pix, float gamma)
 	int k, x, y;
 
 	for (k = 0; k < 256; k++)
-		gamma_map[k] = pow(k / 255.0f, gamma) * 255 + 0.5f;
+		gamma_map[k] = powf(k / 255.0f, gamma) * 255 + 0.5f;
 
 	for (y = 0; y < pix->h; y++)
 	{
